# Conflicts:
#	client/src/main/java/client/scenes/AddExpenseCtrl.java
#	client/src/main/resources/client/scenes/AddExpense.fxml

+++ client/src/main/java/client/scenes/MainCtrl.java
@@ -17,9 +17,11 @@ package client.scenes;
 
 import client.services.I18N;
 import client.services.NotificationHelper;
+import client.services.NotificationService;
 import client.utils.Config;
 import client.utils.User;
 import commons.Event;
+import commons.Expense;
 import commons.Participant;
 import javafx.event.EventHandler;
 import javafx.scene.Parent;
@@ -70,6 +72,7 @@ public class MainCtrl {
 
     private AdminEventsCtrl adminEventsCtrl;
     private Scene adminEvents;
+    private NotificationService notificationService;
 
     public void initialize(Stage primaryStage, Pair<FirstTimeCtrl, Parent> firstTime,
                            Pair<EventOverviewCtrl, Parent> eventOverview,
@@ -110,8 +113,8 @@ public class MainCtrl {
         this.editParticipantCtrl = editParticipant.getKey();
         this.editParticipant = new Scene(editParticipant.getValue());
 
-        this.inviteViewCtrl = inviteView.getKey();
         this.inviteView = new Scene(inviteView.getValue());
+        this.inviteViewCtrl = inviteView.getKey();
 
         this.userSettingsCtrl = userSettings.getKey();
         this.userSettings = new Scene(userSettings.getValue());
@@ -156,6 +159,7 @@ public class MainCtrl {
                 }
             }
         });
+        this.notificationService = new NotificationHelper();
     }
 
     public void showInviteView(Event event){
@@ -278,11 +282,20 @@ public class MainCtrl {
 
     public void showAddExpense(){
         primaryStage.setTitle(I18N.get("window.expense"));
-        addExpenseCtrl.setup(eventOverviewCtrl.getEvent());
+        addExpenseCtrl.setup(eventOverviewCtrl.getEvent(), null);
+        primaryStage.setScene(addExpense);
+    }
+
+    public void showEditExpense(Expense e){
+        primaryStage.setTitle("Splitty: Edit Expense");
+        addExpenseCtrl.setup(eventOverviewCtrl.getEvent(), e);
         primaryStage.setScene(addExpense);
     }
 
     public void deleteAllData(){
+        if (!notificationService.showConfirmation("Delete data", "Are you sure you want to delete your data? This action cannot be undone.")){
+            return;
+        }
         Config.deleteUserConfigFile();
         this.chooseFirstPage(false);
     }
